Initial commit - sets up Employee class

+++ src/main/java/staff/Employee.java
@@ -20,7 +20,7 @@ public abstract class Employee {
         return niNumber;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
